+++ CHANGELOG.md
@@ -19,6 +19,7 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 ### Added
 
 * `restart` command.
+* `noc.config` section.
 
 ### Changed